feat: Improve backend in controller - Divided the router and controllers for code reusability
